requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiogram-dialog==2.1.0
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
APScheduler==3.10.4
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
black==24.2.0
cachetools==5.3.3
certifi==2024.2.2
click==8.1.7
DateTime==5.4
emoji==2.10.1
frozenlist==1.4.1
greenlet==3.0.3
idna==3.6
Jinja2==3.1.3
magic-filter==1.0.12
Mako==1.3.2
MarkupSafe==2.1.5
multidict==6.0.5
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
pydantic==2.5.3
pydantic-settings==2.2.1
pydantic_core==2.14.6
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.2
six==1.16.0
SQLAlchemy==2.0.28
typing_extensions==4.10.0
tzlocal==5.2
yarl==1.9.4
zope.interface==6.2
